refactor: moved gen_flag_prop function
